perfect damage calculation

+++ DiabloII.Items/DiabloII.Items.Reader/Items/ItemProperty.cs
@@ -4,7 +4,7 @@
     {
         // Maybe name should be an enum to be faster to query
         public string Name { get; set; }
-        public int Par { get; set; }
+        public double Par { get; set; }
         public int Minimum { get; set; }
         public int Maximum { get; set; }
         public bool IsPercent { get; set; }
